+++ groups/flamecomics/scrape.js
@@ -20,7 +20,7 @@ export const flameList = async (url = "https://flamecomics.com/") => {
           .find(".bsx .bigor .chapter-list > a:first-child > div > div.epxs")
           .text();
         const chapp = chap.split(" ");
-        const chapter = parseInt(chapp[1]);
+        const chapter = parseFloat(chapp[1]);
 
         const link = $(comic).find(".bsx .bigor .info > a").attr("href");
